requirements.txt
aiofiles==0.8.0
anyio==3.6.2
asgiref==3.4.1
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.0.1
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
croniter==6.0.0
cryptography==40.0.2
dataclasses==0.8
ecdsa==0.19.0
fastapi==0.83.0
fastapi-cli==0.0.1
greenlet==2.0.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.10
immutables==0.19
importlib-metadata==1.7.0
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.9.2
PyMySQL==1.0.2
pyparsing==3.1.4
pytest==7.0.1
pytest-asyncio==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2025.1
rfc3986==1.5.0
rsa==4.9
six==1.17.0
sniffio==1.2.0
SQLAlchemy==1.4.54
starlette==0.19.1
tomli==1.2.3
typing_extensions==4.1.1
uvicorn==0.16.0
zipp==3.6.0
